+++ WebApiTerra1000/Controllers/NomenclatureController.cs
@@ -33,29 +33,22 @@ namespace WebApiTerra1000.Controllers
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/nomenclature/")]
-        public ContentResult GetNomenclature(string code, int id, FormatType format = FormatType.Xml)
-        {
-            return Controller.RunTask(new Task<ContentResult>(() =>
-            {
-                string response = string.IsNullOrEmpty(code)
-                    ? TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclatureFromId, new SqlParameter("id", id))
-                    : TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclatureFromCode, new SqlParameter("code", code));
-                XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
-                XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
-                return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
-            }), format);
-        }
+        public ContentResult GetNomenclature(string code, int id, FormatType format = FormatType.Xml) =>
+            GetNomenclatureWork(SqlQueries.GetNomenclatureFromId, SqlQueries.GetNomenclatureFromCode, code, id, format);
 
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/nomenclature_preview/")]
-        public ContentResult GetNomenclaturePreview(string code, int id, FormatType format = FormatType.Xml)
+        public ContentResult GetNomenclaturePreview(string code, int id, FormatType format = FormatType.Xml) =>
+            GetNomenclatureWork(SqlQueries.GetNomenclatureFromIdPreview, SqlQueries.GetNomenclatureFromCodePreview, code, id, format);
+
+        private ContentResult GetNomenclatureWork(string urlId, string urlCode, string code, int id, FormatType format = FormatType.Xml)
         {
             return Controller.RunTask(new Task<ContentResult>(() =>
             {
                 string response = string.IsNullOrEmpty(code)
-                    ? TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclatureFromIdPreview, new SqlParameter("id", id))
-                    : TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclatureFromCodePreview, new SqlParameter("code", code));
+                    ? TerraUtils.Sql.GetResponse<string>(SessionFactory, urlId, new SqlParameter("id", id))
+                    : TerraUtils.Sql.GetResponse<string>(SessionFactory, urlCode, new SqlParameter("code", code));
                 XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
                 XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
                 return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
@@ -66,27 +59,22 @@ namespace WebApiTerra1000.Controllers
         [HttpGet()]
         [Route("api/nomenclatures/")]
         public ContentResult GetNomenclatures(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
-            FormatType format = FormatType.Xml)
-        {
-            return Controller.RunTask(new Task<ContentResult>(() =>
-            {
-                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclatures,
-                    TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
-                XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
-                XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
-                return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
-            }), format);
-        }
+            FormatType format = FormatType.Xml) =>
+            GetNomenclaturesWork(SqlQueries.GetNomenclatures, startDate, endDate, offset, rowCount, format);
 
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/nomenclatures_preview/")]
         public ContentResult GetNomenclaturesPreview(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
+            FormatType format = FormatType.Xml) =>
+            GetNomenclaturesWork(SqlQueries.GetNomenclaturesPreview, startDate, endDate, offset, rowCount, format);
+
+        private ContentResult GetNomenclaturesWork(string url, DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
             FormatType format = FormatType.Xml)
         {
             return Controller.RunTask(new Task<ContentResult>(() =>
             {
-                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclaturesPreview,
+                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, url,
                     TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
                 XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
                 XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
@@ -97,26 +85,23 @@ namespace WebApiTerra1000.Controllers
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/nomenclaturescosts/")]
-        public ContentResult GetNomenclaturesCosts(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10, FormatType format = FormatType.Xml)
-        {
-            return Controller.RunTask(new Task<ContentResult>(() =>
-            {
-                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclaturesCosts,
-                    TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
-                XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
-                XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
-                return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
-            }), format);
-        }
+        public ContentResult GetNomenclaturesCosts(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
+            FormatType format = FormatType.Xml) => 
+            GetNomenclaturesCostsWork(SqlQueries.GetNomenclaturesCosts, startDate, endDate, offset, rowCount, format);
 
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/nomenclaturescosts_preview/")]
-        public ContentResult GetNomenclaturesCostsPreview(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10, FormatType format = FormatType.Xml)
+        public ContentResult GetNomenclaturesCostspreview(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
+            FormatType format = FormatType.Xml) => 
+            GetNomenclaturesCostsWork(SqlQueries.GetNomenclaturesCostsPreview, startDate, endDate, offset, rowCount, format);
+
+        private ContentResult GetNomenclaturesCostsWork(string url, DateTime startDate, DateTime endDate, int offset = 0, 
+            int rowCount = 10, FormatType format = FormatType.Xml)
         {
             return Controller.RunTask(new Task<ContentResult>(() =>
             {
-                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclaturesCostsPreview,
+                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, url,
                     TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
                 XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
                 XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));