+++ WebApiTerra1000/Utils/SqlQueries.cs
@@ -33,12 +33,24 @@ SELECT [dbo].[fnGetException123]() [fnGetException]
 SELECT[IIS].[fnGetDeliveryPlaces] (:StartDate, :EndDate, :Offset, :RowCount)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
 
+        public static string GetDeliveryPlacesPreview => @"
+SELECT[IIS].[fnGetDeliveryPlacesPreview] (:StartDate, :EndDate, :Offset, :RowCount)
+            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
+
         public static string GetContragent => @"
 SELECT [IIS].[fnGetContragentByID] (:ID)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
 
+        public static string GetContragentPreview => @"
+SELECT [IIS].[fnGetContragentByIDPreview] (:ID)
+            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
+
         public static string GetContragents => @"
 SELECT[IIS].[fnGetContragentChangesList] (:StartDate, :EndDate, :Offset, :RowCount)
+            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
+
+        public static string GetContragentsPreview => @"
+SELECT[IIS].[fnGetContragentChangesListPreview] (:StartDate, :EndDate, :Offset, :RowCount)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
 
         public static string GetNomenclatureFromCode => @"
@@ -77,12 +89,25 @@ SELECT [IIS].[fnGetNomenclatureListPreview] (:StartDate, :EndDate, :Offset, :Row
 SELECT [IIS].[GetRefShipmentsById] (:ID)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
 
+        public static string GetShipmentPreview => @"
+SELECT [IIS].[GetRefShipmentsByIdPreview] (:ID)
+            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
+
         public static string GetShipments => @"
 SELECT [IIS].[GetRefShipmentsByDocDate] (:StartDate,:EndDate,:Offset,:RowCount)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
 
+        public static string GetShipmentsPreview => @"
+SELECT [IIS].[GetRefShipmentsByDocDatePreview] (:StartDate,:EndDate,:Offset,:RowCount)
+            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
+
         public static string GetSummary => @"
 SELECT [IIS].[fnGetSummaryList] (:StartDate, :EndDate)
+            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
+
+
+        public static string GetSummaryPreview => @"
+SELECT [IIS].[fnGetSummaryListPreview] (:StartDate, :EndDate)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
 
     }