+++ WebApiTerra1000Tests/TestsUtils.cs
@@ -1,8 +1,6 @@
 // This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-using DataCore.DAL.Models;
-using Microsoft.Data.SqlClient;
 using NUnit.Framework;
 using System.Runtime.CompilerServices;
 
@@ -16,38 +14,20 @@ namespace DataCoreTests
         #region Public and private fields and properties
 
         public static bool IsPrepare { get; private set; } = false;
-        public static DataAccessEntity DataAccess { get; private set; } = null;
-        public static SqlConnection SqlCon { get; private set; }
+        public static string GetUrlDev { get; private set; } = "https://t1000-dev.kolbasa-vs.local:443/api/";
+        public static string GetUrlDevPreview { get; private set; } = "https://t1000-dev-preview.kolbasa-vs.local:443/api/";
+        public static string GetUrlProd { get; private set; } = "https://t1000.kolbasa-vs.local:443/api/";
+        public static string GetUrlProdPreview { get; private set; } = "https://t1000-preview.kolbasa-vs.local:443/api/";
 
         #endregion
 
         #region Public and private methods
-
-        public static void SqlPrepare()
-        {
-            if (IsPrepare) return;
-
-            //IConfiguration config = new ConfigurationBuilder()
-            //    .AddJsonFile("appsettings.json")
-            //    .AddEnvironmentVariables()
-            //    .Build();
-            //JsonSettingsEntity jsonSettings = new(config);
-
-            //JsonSettingsConfigEntity jsonSettings = new("");
-            //DataAccess = new(jsonSettings);
-
-            //SqlCon = SqlConnectFactory.GetConnection(DataAccess.GetSession().Connection.ConnectionString);
-
-            IsPrepare = true;
-        }
 
         public static void MethodStart([CallerMemberName] string memberName = "")
         {
             TestContext.WriteLine(@"--------------------------------------------------------------------------------");
             TestContext.WriteLine($@"{memberName} start.");
             TestContext.WriteLine();
-
-            SqlPrepare();
         }
 
         public static void MethodComplete([CallerMemberName] string memberName = "")